requirements.txt
langchain
pypdf
langchain-community
psycopg2
python-dotenv
numpy
pgvector
